refactor: a single integration test file for all triumph source events
