+++ Api/tests/functional/Triumph/Event/TriumphSourceEventCest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Mush\Tests\functional\Triumph\Event;
 
 use Mush\Action\Enum\ActionEnum;
+use Mush\Daedalus\Event\DaedalusCycleEvent;
 use Mush\Daedalus\Event\DaedalusEvent;
 use Mush\Game\Service\EventServiceInterface;
 use Mush\Tests\AbstractFunctionalTest;
@@ -13,15 +14,30 @@ use Mush\Tests\FunctionalTester;
 /**
  * @internal
  */
-final class DaedalusEventCest extends AbstractFunctionalTest
+final class TriumphSourceEventCest extends AbstractFunctionalTest
 {
     private EventServiceInterface $eventService;
 
-    public function _before(FunctionalTester $I)
+    public function _before(FunctionalTester $I): void
     {
         parent::_before($I);
 
         $this->eventService = $I->grabService(EventServiceInterface::class);
+    }
+
+    public function shouldGiveTriumphOnDaedalusNewCycleEvent(FunctionalTester $I): void
+    {
+        $this->player->setTriumph(0);
+
+        $event = new DaedalusCycleEvent(
+            daedalus: $this->daedalus,
+            tags: [],
+            time: new \DateTime(),
+        );
+        $this->eventService->callEvent($event, DaedalusCycleEvent::DAEDALUS_NEW_CYCLE);
+
+        // human cyclic triumph
+        $I->assertEquals(1, $this->player->getTriumph());
     }
 
     public function shouldGiveTriumphOnDaedalusEvent(FunctionalTester $I): void
@@ -35,6 +51,7 @@ final class DaedalusEventCest extends AbstractFunctionalTest
         );
         $this->eventService->callEvent($event, DaedalusEvent::FINISH_DAEDALUS);
 
+        // return to sol human triumph
         $I->assertEquals(20, $this->player->getTriumph());
     }
 }
